requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.17
botocore==1.19.17
certifi==2020.11.8
chardet==3.0.4
future==0.18.2
idna==2.10
imutils==0.5.3
jmespath==0.10.0
meraki==1.3.0
multidict==5.0.0
numpy==1.18.2
opencv-python==4.2.0.34
paho-mqtt==1.5.0
Pillow==8.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.25.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
six==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
webexteamssdk==1.6
yarl==1.6.2
